Fix user priming
