+++ packages/common/src/api/tan-query/utils/formatTrackData.ts
@@ -1,0 +1,74 @@
+import { Track, TrackMetadata } from '~/models/Track'
+
+/**
+ * Add the cosigned status to the track
+ */
+const setIsCoSigned = <T extends TrackMetadata>(track: T) => {
+  const { remix_of } = track
+
+  const remixOfTrack = remix_of?.tracks?.[0]
+
+  const isCoSigned =
+    remixOfTrack &&
+    (remixOfTrack.has_remix_author_saved ||
+      remixOfTrack.has_remix_author_reposted)
+
+  if (isCoSigned) {
+    return {
+      ...track,
+      _co_sign: remix_of!.tracks[0]
+    }
+  }
+  return track
+}
+
+/**
+ * When a track is not unlisted, even if field visibility is set
+ * we should coerce the track into a state where socials are visible.
+ * @param track
+ * @returns track with repaired field visibility
+ */
+const setFieldVisibility = <T extends TrackMetadata>(track: T) => {
+  const { is_unlisted } = track
+  if (!is_unlisted) {
+    // Public track
+    return {
+      ...track,
+      field_visibility: {
+        ...track.field_visibility,
+        genre: true,
+        mood: true,
+        tags: true,
+        share: true,
+        play_count: true,
+        remixes: true
+      }
+    }
+  }
+  return track
+}
+
+/**
+ * NOTE: This is a temporary fix for a backend bug: The field followee_saves is not defined.
+ * This is a stopgap to prevent the client from erroring and should be removed after fixed.
+ * The current erroneous disprov endpoint is `/feed/reposts/<userid>`
+ * @param track
+ */
+const setDefaultFolloweeSaves = <T extends TrackMetadata>(track: T) => {
+  return {
+    ...track,
+    followee_saves: track?.followee_saves ?? []
+  }
+}
+
+/**
+ * Reformats a track to be used internally within the client
+ * This method should *always* be called before a track is cached.
+ */
+export const formatTrackData = <T extends TrackMetadata>(track: T): Track => {
+  const withCosign = setIsCoSigned(track)
+  const withFieldVisibility = setFieldVisibility(withCosign)
+
+  const withDefaultSaves = setDefaultFolloweeSaves(withFieldVisibility)
+  return withDefaultSaves
+}
